fix: structure directories
