+++ src/models/steps/model_loading/common/MiniGPT/minigpt_model_context_loader.py
@@ -1,0 +1,7 @@
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+def minigpt_load_model(weights_path_to_load: str, device: str):
+    pass
