+++ src/models/steps/model_loading/common/ultralytics/ultralytics_model_context_loader.py
