+++ src/models/steps/model_training/common/ultralytics_model_context_trainer.py
