+++ src/steps/data_extraction/processing/processing_data_extractor.py
@@ -1,11 +1,16 @@
 import os
-from typing import Optional
+from typing import Optional, Union
 
 from src import Pipeline
 from src import step
+from src.models.contexts.processing.picsellia_datalake_processing_context import (
+    PicselliaDatalakeProcessingContext,
+)
 from src.models.contexts.processing.picsellia_processing_context import (
     PicselliaProcessingContext,
 )
+from src.models.dataset.processing.datalake_collection import DatalakeCollection
+from src.models.dataset.processing.datalake_context import DatalakeContext
 from src.models.dataset.common.dataset_context import DatasetContext
 from src.models.dataset.processing.processing_dataset_collection import (
     ProcessingDatasetCollection,
@@ -94,3 +99,32 @@ def processing_dataset_collection_extractor() -> ProcessingDatasetCollection:
     )
     dataset_collection.download_all()
     return dataset_collection
+
+
+@step
+def processing_datalake_extractor() -> Union[DatalakeContext, DatalakeCollection]:
+    context: PicselliaDatalakeProcessingContext = Pipeline.get_active_context()
+    destination_path = get_destination_path(context.job_id)
+    input_datalake_context = DatalakeContext(
+        datalake_name="input",
+        datalake=context.input_datalake,
+        destination_path=destination_path,
+        data_ids=context.data_ids,
+        use_id=context.use_id,
+    )
+    if context.output_datalake:
+        output_datalake_context = DatalakeContext(
+            datalake_name="output",
+            datalake=context.output_datalake,
+            destination_path=destination_path,
+            use_id=context.use_id,
+        )
+        datalake_collection = DatalakeCollection(
+            input_datalake_context=input_datalake_context,
+            output_datalake_context=output_datalake_context,
+        )
+        datalake_collection.download_all()
+        return datalake_collection
+    else:
+        input_datalake_context.download_data(image_dir=input_datalake_context.image_dir)
+        return input_datalake_context
